+++ screens/createCircle/CircleCreate.jsx
@@ -5,20 +5,14 @@ import { Image } from 'expo-image';
 import { BottomModal } from '../../components/Modal';
 import { styles } from '../Profile/styles';
 import Checkbox from 'expo-checkbox';
-import {
-  useCameraPermissions,
-  useMediaLibraryPermissions,
-  PermissionStatus,
-  MediaTypeOptions,
-  launchImageLibraryAsync,
-  launchCameraAsync,
-} from 'expo-image-picker';
 import { useRecoilState } from 'recoil';
 import { newCircleState } from '../../stores/circle-store';
 import { ScrollView } from 'react-native';
 import instance, { circleInstance } from '../../api/instance';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { createCircle } from '../../api/circleApi';
+import useCamera from '../../hooks/useCamera';
+import useMediaLibrary from '../../hooks/useMediaLibrary';
 
 const CircleCreate = () => {
   const queryClient = useQueryClient();
@@ -31,99 +25,22 @@ const CircleCreate = () => {
   const [newCircle, setNewCircle] = useRecoilState(newCircleState);
   const [isModalVisible, setModalVisible] = useState(false);
 
-  const [imagePermissionInformation, requestImagePermission] = useMediaLibraryPermissions(); // 미디어 라이브러리 접근 권한
-  const [cameraPermissionInformation, requestCameraPermission] = useCameraPermissions(); // 카메라 접근 권한
+  const { selectImageHandler } = useMediaLibrary(onImageCaptured);
+  const { takeImageHandler } = useCamera(onImageCaptured);
 
   const navigation = useNavigation();
 
-  async function verifyImagePermissions() {
-    if (imagePermissionInformation.status !== PermissionStatus.GRANTED) {
-      const permissionResponse = await requestImagePermission();
-      return permissionResponse.granted;
-    }
+  const onImageCaptured = uri => {
+    setNewCircle({ ...newCircle, image: uri });
+    setModalVisible(!isModalVisible);
+  };
 
-    if (imagePermissionInformation.status === PermissionStatus.DENIED) {
-      Alert.alert('미디어 라이브러리 권한이 필요합니다.', '설정에서 권한을 허용해주세요.', [
-        { text: '취소', style: 'cancel' },
-        { text: '설정으로 이동', onPress: () => Linking.openSettings() },
-      ]);
-      return false;
-    }
-
-    return true;
-  }
-
-  async function verifyCameraPermissions() {
-    if (cameraPermissionInformation.status !== PermissionStatus.GRANTED) {
-      const permissionResponse = await requestCameraPermission();
-      return permissionResponse.granted;
-    }
-
-    if (cameraPermissionInformation.status === PermissionStatus.DENIED) {
-      Alert.alert('카메라 권한이 필요합니다.', '설정에서 권한을 허용해주세요.', [
-        { text: '취소', style: 'cancel' },
-        { text: '설정으로 이동', onPress: () => Linking.openSettings() },
-      ]);
-      return false;
-    }
-
-    return true;
-  }
-
-  async function selectImageHandler() {
-    try {
-      const hasPermission = await verifyImagePermissions();
-
-      if (!hasPermission) {
-        return;
-      }
-
-      const image = await launchImageLibraryAsync({
-        mediaTypes: MediaTypeOptions.Images,
-        allowsEditing: true,
-        aspect: [1, 1],
-        quality: 0.5,
-      });
-
-      if (!image.canceled) {
-        setNewCircle({ ...newCircle, image: image.assets[0].uri });
-        setModalVisible(!isModalVisible);
-      }
-    } catch (error) {
-      Alert.alert('미디어 라이브러리를 사용할 수 없습니다.', '다시 시도해주세요.');
-      console.log(error);
-    }
-  }
-
-  async function takeImageHandler() {
-    try {
-      const hasPermission = await verifyCameraPermissions();
-
-      if (!hasPermission) {
-        return;
-      }
-
-      const image = await launchCameraAsync({
-        allowsEditing: true,
-        aspect: [16, 9],
-        quality: 0.5,
-      });
-      if (!image.canceled) {
-        setNewCircle({ ...newCircle, image: image.assets[0].uri });
-        setModalVisible(!isModalVisible);
-      }
-      // console.log(image);
-    } catch (error) {
-      Alert.alert('카메라를 사용할 수 없습니다.', '다시 시도해주세요.');
-    }
-  }
-
-  function deleteImageHandler() {
+  const onDeleteThumbnail = () => {
     setNewCircle({ ...newCircle, image: null });
     setModalVisible(!isModalVisible);
-  }
+  };
 
-  const toggleModal = () => {
+  const onToggleModal = () => {
     setModalVisible(!isModalVisible);
   };
   const onPressConfirm = () => {
@@ -162,10 +79,10 @@ const CircleCreate = () => {
         icon: require('../../assets/icons/trash.png'),
         iconStyle: styles.trash,
         textStyle: { color: 'red' }, // 빨간색 텍스트
-        onPress: deleteImageHandler,
+        onPress: onDeleteThumbnail,
       },
     ],
-    [selectImageHandler, takeImageHandler, deleteImageHandler],
+    [selectImageHandler, takeImageHandler, onDeleteThumbnail],
   );
 
   return (
@@ -173,7 +90,7 @@ const CircleCreate = () => {
       style={{ flex: 1, backgroundColor: '#FFF' }}
       contentContainerStyle={{ alignItems: 'center', paddingBottom: 25 }}
       showsVerticalScrollIndicator={false}>
-      <BottomModal isModalVisible={isModalVisible} toggleModal={toggleModal} buttons={editButtons} />
+      <BottomModal isModalVisible={isModalVisible} onToggleModal={onToggleModal} buttons={editButtons} />
       {newCircle.image ? (
         <Image source={newCircle.image} style={[styles.image, { borderRadius: 20 }]} contentFit="cover" />
       ) : (
@@ -185,7 +102,7 @@ const CircleCreate = () => {
           />
         </View>
       )}
-      <Pressable style={styles.pinkButton} onPress={toggleModal}>
+      <Pressable style={styles.pinkButton} onPress={onToggleModal}>
         <Text style={styles.buttonText}>사진 등록</Text>
       </Pressable>
       <Pressable onPress={onPressCircleName}>
